// src/app/models/admin/dangkyhoatdongkhoangsan/dangkythamdo/dkthamdotoadokhuvuc.model.ts
export class OutputDkThamDoToaDoKhuVucModel {
    idthamdotoadokhuvuc: string;
    sohieu: string;
    toadox: string;
    toadoy: string;
    iddangkythamdo: string;
    idthamdokhuvuc: string;
    loaikhuvuc: number;
    loaicapphep: number;
    siteid: string;
    thutu: number
}

export class InputDkThamDoToaDoKhuVucModel {
    idthamdotoadokhuvuc: string;
    sohieu: string;
    toadox: string;
    toadoy: string;
    iddangkythamdo: string;
    idthamdokhuvuc: string;
    loaicapphep: number;
    loaikhuvuc: number;
    thutu: number
}